Fix build. Move into a separate component ECR-42
