+++ Hardware/ESPIDF-Server/components/ble_service/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "src/ble_communication.cpp"
+        INCLUDE_DIRS "include"
+        PRIV_REQUIRES wifi_service bt nvs_flash)
